requirements.txt
crewai
langchain-google-genai
openpyxl
streamlit>=1.28.0
PyYAML
keyring>=24.0.0
streamlit-cookies-manager>=0.2.0
pytest>=7.0.0
